changed workspace, solution. project names
