+++ MyMate_Client/ClientModules/Classes/MdlChatroom.cs
@@ -4,12 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Jin.Classes
+namespace ClientModules.Classes
 {
-    public class MdlMessage
+    public class MdlChatroom
     {
-        public int? UserCode { get; set; }
-        public int? ChannelCode { get; set; }
+        public int? ChatroomCode { get; set; }
         public int? ServerCode { get; set; }
         public string? Message { get; set; }
         public string? Date { get; set; }
