+++ MyMate_Client/ClientModules/ClientModules.csproj
