+++ MyMate_Client/MyMate_Client.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32901.215
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientModules", "JinConsole\ClientModules.csproj", "{EEC822C4-FE57-4F23-822A-E96CCE33C266}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Protocol", "..\MyMate_Network\MyMate_Network\Protocal\Protocol.csproj", "{C4234080-C265-41B6-B899-914366307188}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientNetwork", "..\MyMate_Network\MyMate_Network\ClientNetwork\ClientNetwork.csproj", "{BE049136-128F-4E9E-936B-A9DB5EB9543C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientModules", "ClientModules\ClientModules.csproj", "{3A2F0545-7226-4962-832A-AC7E848CA428}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EEC822C4-FE57-4F23-822A-E96CCE33C266}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EEC822C4-FE57-4F23-822A-E96CCE33C266}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EEC822C4-FE57-4F23-822A-E96CCE33C266}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EEC822C4-FE57-4F23-822A-E96CCE33C266}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C4234080-C265-41B6-B899-914366307188}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C4234080-C265-41B6-B899-914366307188}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C4234080-C265-41B6-B899-914366307188}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +23,10 @@ Global
 		{BE049136-128F-4E9E-936B-A9DB5EB9543C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BE049136-128F-4E9E-936B-A9DB5EB9543C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BE049136-128F-4E9E-936B-A9DB5EB9543C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3A2F0545-7226-4962-832A-AC7E848CA428}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3A2F0545-7226-4962-832A-AC7E848CA428}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3A2F0545-7226-4962-832A-AC7E848CA428}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3A2F0545-7226-4962-832A-AC7E848CA428}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
